Separate reporter config
